backend/requirements.txt
anyio==4.2.0
bardapi==0.1.39
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
